ec2 tabpane
